working + normed (but probably a lot of bugs imo)

+++ libft/ft_putnbr_fd.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	ft_putnbr_fd(int nb, int fd)
+void	ft_putnbr_fd(long long int nb, int fd)
 {
 	int	to_print;
